**WIP********** Show Ambient notif. when new track is played even if AD is off [1/2]
This let us trigger an Ambient Display notification even if it's
disabled in Display-Ambient settings, so we can know the new track
title/artist.
This is handy when the player starts playing a new track or even
when we use volume rocker skip track feature when screen is off.

We use provided Doze services to skip this when the proximity
sensor is covered, like other ambient modes.

Change-Id: I1365d3a7ca6a03d369584bfcb8a1b421b2015b7a